+++ src/redux/actions/cities.actions.js
@@ -1,6 +1,4 @@
-export const types = {
-    DELETE_LIST_ITEM: 'DELETE_LIST_ITEM',
-};
+import { types } from '../constants/cities.types';
 
 export const actions = {
     deleteListItem: (id) => ({ type: types.DELETE_LIST_ITEM, id, }),
