+++ src/redux/constants/cities.types.js
@@ -1,0 +1,3 @@
+export const types = {
+    DELETE_LIST_ITEM: 'DELETE_LIST_ITEM',
+};
