+++ src/redux/constants/city.types.js
@@ -1,0 +1,5 @@
+export const types = {
+    FETCH_FORECAST_REQUEST: 'FETCH_FORECAST_REQUEST',
+    FETCH_FORECAST_SUCCESS: 'FETCH_FORECAST_SUCCESS',
+    FETCH_FORECAST_FAILURE: 'FETCH_FORECAST_FAILURE',
+};
